Previous commit requiredness changes broke testValueIRParser temporarily. This also moves checkedConvertFrom up to PContainer
